Refactor gwt.py into modular gwtlib package structure
Transform monolithic 1,449-line gwt.py into clean modular architecture:

- Reduce gwt.py to 45-line thin shim that preserves PEP 723 header
  and re-exports public API for backward compatibility
- Create gwtlib/ package with 11 focused modules organized by concern:
  * config.py - Configuration and HAS_TOML centralization
  * git_ops.py - Standardized git command wrappers
  * paths.py - Path utilities for worktrees and display
  * branches.py - Branch detection and remote operations
  * parsing.py - Git worktree list parsing and reconciliation
  * display.py - Formatting and listing functions
  * worktrees.py - Worktree creation, switching, and removal
  * resolution.py - Git directory auto-detection
  * cli.py - CLI argument parsing and dispatch
  * api.py - Public API aggregator for test compatibility
- Add validation tests for module imports and API re-exports
- Standardize all git subprocess calls through git_ops wrappers
- Maintain 100% backward compatibility (all 23 tests pass)

All modules stay well under 800-line target (largest: 334 lines).
Preserves original behavior while improving maintainability.
