+++ gwtlib/__init__.py
@@ -1,0 +1,3 @@
+# gwtlib/__init__.py
+__all__ = []
+__version__ = "0.0.0-refactor-split"
